# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/apis": {
            "get": {
                "description": "查询apis信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API相关"
                ],
                "summary": "获取apis信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "apiGroup",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "新增不用传ID，修改才传ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API相关"
                ],
                "summary": "新增/修改API信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "apiGroup",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "必须大写",
                        "name": "method",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "如果API为/user就/user，不要写成/user/",
                        "name": "path",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除apis",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API相关"
                ],
                "summary": "删除apis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "传新增/修改API的参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/apis/group": {
            "get": {
                "description": "获取API所有组",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API相关"
                ],
                "summary": "获取API所有组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
        "/buttons": {
            "get": {
                "description": "查询所有/指定条件按钮的信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "按钮相关"
                ],
                "summary": "查询按钮信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "buttonCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "buttonDesc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "修改才需要传，没有传算新增",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "menuId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "新增不用传ID，修改才传ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "按钮相关"
                ],
                "summary": "新增/修改按钮信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "传新增/修改按钮的参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateButtonsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定按钮",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "按钮相关"
                ],
                "summary": "删除按钮",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "按钮ID",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.IdsReq": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "api.UpdateButtonReq": {
            "type": "object",
            "required": [
                "buttonCode",
                "menuId"
            ],
            "properties": {
                "buttonCode": {
                    "type": "string"
                },
                "buttonDesc": {
                    "type": "string"
                },
                "menuId": {
                    "type": "integer"
                }
            }
        },
        "api.UpdateButtonsReq": {
            "type": "object",
            "required": [
                "buttons"
            ],
            "properties": {
                "buttons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.UpdateButtonReq"
                    }
                }
            }
        },
...
    }
}